requirements.txt
discord.py==2.3.2
edge-tts
emoji
python-dotenv
